requirements.txt
asgiref==3.8.1
boto3==1.28.53
botocore==1.31.53
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.0.1
crispy-bootstrap5==2024.10
cryptography==39.0.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==5.1.2
django-allauth==65.0.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-honeypot==1.2.1
django-import-export==3.3.1
django-recaptcha==4.1.0
django-storages==1.14
django-imagekit==6.0.0
et_xmlfile==2.0.0
fastapi==0.135.3
gunicorn==23.0.0
idna==3.4
jmespath==1.0.1
mailchimp-marketing==3.0.80
MarkupPy==1.14
oauthlib==3.3.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==23.1
pillow==11.2.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.6.2
six==1.16.0
smtp2go==2.0.0
smtp2go-django==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.4.3
tablib==3.5.0
typing_extensions==4.9.0
urllib3==1.26.14
uvicorn==0.44.0
whitenoise==6.7.0
xlrd==2.0.1
xlwt==1.3.0
